requirements.txt
systemd-python
fastapi
tomli
requests
websockets
pyamdgpuinfo
wakeonlan
psutil
uvicorn
hurry.filesize
